api/requirements.txt
fastapi==0.100.0
accelerate==1.2.1
psycopg2-binary
boto3==1.26.8
uvicorn==0.22.0
pydantic==2.10.5
mlflow==2.19.0
transformers==4.48.0
pandas==1.5.3
